pyproject.toml
[project]
name = "guess_explainr"
version = "0.1"
description = "Geography learning app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-cookie3>=0.20",
    "frozendict>=2.4.7",
    "geoguessr-async>=2.1.0",
    "geopy>=2.4.1",
    "litestar>=2.14",
    "markdown>=3.10.2",
    "niquests>=3.18.6",
    "platformdirs>=4.9.6",
    "pydantic-ai>=1.86.0",
    "streetview>=0.1.0",
    "streetview-dl",
    "uvicorn[standard]>=0.45",
]

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pyrefly>=0.16",
    "pytest>=8.3",
    "pytest-asyncio>=0.26",
    "ruff>=0.11",
    "requests>=2.33.1",
    "beautifulsoup4>=4.14.3",
    "tqdm>=4.67.3",
    "pillow>=12.2.0",
]

[project.urls]
sourcecode = "https://codeberg.org/atollk/guess_explainr/"
GitHub = "https://github.com/atollk/guess_explainr/"

[project.scripts]
guess-explainr = "guess_explainr.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guess_explainr"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyrefly]
project_includes = ["src"]
python_version = "3.12"
search_path = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
# streetview pulls in jupyter for no reason
exclude-dependencies = ["jupyter"]
